feat: [issues3-logical-move-workstation-config-001] - LOGICAL_MOVE draft validation skips worker and prompt requirements
