+++ ui/src/features/current-selection/workstation-selection/lib/editable-workstation-configuration-validation.ts
@@ -1,0 +1,150 @@
+import {
+  workerSupportsPollerBehavior,
+  workstationBehaviorRequiresPrompt,
+} from "../../../current-factory-definition/lib/workstation-behavior";
+import type {
+  EditableWorkstationDraft,
+  resolveEditableWorkstationValues,
+} from "../../../current-factory-definition/lib/workstation-editable-values";
+import { workstationRequiresWorkerAssignment } from "../../../current-factory-definition/lib/workstation-worker-assignment";
+import type {
+  EditableWorkstationPromptValidationState,
+  EditableWorkstationValidationErrors,
+  EditableWorkstationWorkerOptionsState,
+} from "./detail-card-types";
+import {
+  getWorkstationDetailMessages,
+  type WorkstationDetailMessages,
+} from "../messages/workstation-detail";
+
+export function validateEditableWorkstationDraft(
+  draft: EditableWorkstationDraft,
+  selectedEditableValues?: ReturnType<typeof resolveEditableWorkstationValues>,
+  promptValidationState: EditableWorkstationPromptValidationState = {
+    status: "idle",
+  },
+  messages: Pick<
+    WorkstationDetailMessages,
+    | "editableConfigurationPromptRequired"
+    | "editableConfigurationPromptValidationLoading"
+    | "editableConfigurationPromptValidationErrorPrefix"
+    | "editableConfigurationPromptFieldHint"
+    | "editableConfigurationBehaviorPollerWorkerUnsupported"
+    | "editableConfigurationWorkerRequired"
+    | "editableConfigurationWorkerUnavailable"
+  > = getWorkstationDetailMessages(undefined),
+): EditableWorkstationValidationErrors {
+  const validationErrors: EditableWorkstationValidationErrors = {};
+  const requiresWorkerAssignment =
+    selectedEditableValues == null ||
+    workstationRequiresWorkerAssignment({
+      type: selectedEditableValues.workstationType,
+    });
+  const promptIsRequired =
+    requiresWorkerAssignment &&
+    workstationBehaviorRequiresPrompt(draft.behavior);
+
+  if (requiresWorkerAssignment) {
+    if (draft.workerName.trim().length === 0) {
+      validationErrors.workerName =
+        messages.editableConfigurationWorkerRequired;
+    } else if (
+      selectedEditableValues &&
+      !selectedEditableValues.workerOptions.includes(draft.workerName)
+    ) {
+      validationErrors.workerName =
+        messages.editableConfigurationWorkerUnavailable;
+    }
+    if (
+      draft.behavior === "POLLER" &&
+      selectedEditableValues &&
+      !workerSupportsPollerBehavior(
+        draft.workerName.trim().length === 0
+          ? null
+          : {
+              type: selectedEditableValues.workerTypeByName[draft.workerName],
+            },
+      )
+    ) {
+      validationErrors.behavior =
+        messages.editableConfigurationBehaviorPollerWorkerUnsupported;
+    }
+  }
+
+  if (promptIsRequired && draft.prompt.trim().length === 0) {
+    validationErrors.prompt = messages.editableConfigurationPromptRequired;
+  } else if (promptIsRequired && promptValidationState.status === "loading") {
+    validationErrors.prompt =
+      messages.editableConfigurationPromptValidationLoading;
+  } else if (promptIsRequired && promptValidationState.status === "error") {
+    validationErrors.prompt = `${messages.editableConfigurationPromptValidationErrorPrefix} ${promptValidationState.errorMessage}`;
+  } else if (
+    promptIsRequired &&
+    promptValidationState.status === "ready" &&
+    (!promptValidationState.result.valid ||
+      promptValidationState.diagnostics.length > 0)
+  ) {
+    validationErrors.prompt = messages.editableConfigurationPromptFieldHint;
+  }
+
+  return validationErrors;
+}
+
+export function hasEditableWorkstationValidationErrors(
+  validationErrors: EditableWorkstationValidationErrors,
+): boolean {
+  return Boolean(
+    validationErrors.behavior ||
+      validationErrors.prompt ||
+      validationErrors.runnerName ||
+      validationErrors.workerName,
+  );
+}
+
+export function resolveWorkerOptionsState(
+  draft: EditableWorkstationDraft,
+  selectedEditableValues: ReturnType<typeof resolveEditableWorkstationValues>,
+  messages: Pick<
+    WorkstationDetailMessages,
+    | "editableConfigurationEmpty"
+    | "editableConfigurationWorkerMissing"
+    | "editableConfigurationWorkerOptionsEmpty"
+  >,
+): EditableWorkstationWorkerOptionsState {
+  if (!selectedEditableValues) {
+    return {
+      message: messages.editableConfigurationEmpty,
+      status: "error",
+    };
+  }
+
+  if (
+    !workstationRequiresWorkerAssignment({
+      type: selectedEditableValues.workstationType,
+    })
+  ) {
+    return {
+      options: selectedEditableValues.workerOptions,
+      status: "ready",
+    };
+  }
+
+  if (selectedEditableValues.workerOptions.length === 0) {
+    return {
+      message: messages.editableConfigurationWorkerOptionsEmpty,
+      status: "empty",
+    };
+  }
+
+  if (!selectedEditableValues.workerOptions.includes(draft.workerName)) {
+    return {
+      message: messages.editableConfigurationWorkerMissing,
+      status: "error",
+    };
+  }
+
+  return {
+    options: selectedEditableValues.workerOptions,
+    status: "ready",
+  };
+}
